ref: Merge ReplayClipPreviewWrapper into GroupReplaysTableInner where its used
